% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_utf8}
\alias{convert_utf8}
\title{Convert utf8}
\usage{
convert_utf8(x)
}
\arguments{
\item{x}{character class vector.}
}
\description{
The \code{convert_utf8()} convert utf8 to unicode character.
}
\examples{
convert_utf8('\u00f3')
convert_utf8('\u00e3o')
}
