% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{detect_file}
\alias{detect_file}
\title{Get the files list in directory}
\usage{
detect_file(path, measure = "mb", format_column = TRUE)
}
\arguments{
\item{path}{character class vector that indicates the path from file list.}

\item{measure}{character class vector that indicates the unit of measure for calculating the object's size in the calling environment. Option is 'bytes', 'kb, 'mb', 'gb' and 'tb'.}

\item{format_column}{logical class vector to format column with characters of length > 16.}
}
\description{
The \code{detect_file()} detect the files with extension in directory.
}
\examples{
detect_file(getwd())
}
