% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_accent}
\alias{extract_accent}
\title{Extract the accent type from the vector in the data frame}
\usage{
extract_accent(data, format_column = TRUE)
}
\arguments{
\item{data}{object class data frame.}

\item{format_column}{logical class vector to format column with characters of length > 16.}
}
\description{
The \code{extract_accent()} extract the accent type from the vector in the data frame.
}
\examples{
data_i <- apply(stadium, 2, convert_utf8)
data_i <- as.data.frame(data_i)
identify::extract_accent(data_i)
}
