% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_punctuation}
\alias{extract_punctuation}
\title{Extract the punctuation type from the vector in the data frame}
\usage{
extract_punctuation(data, format_column = TRUE)
}
\arguments{
\item{data}{object class data frame.}

\item{format_column}{logical class vector to format column with characters of length > 16.}
}
\description{
The \code{extract_punctuation()} extract the punctuation type from the vector in the data frame.
}
\examples{
extract_punctuation(identify::stadium)
extract_punctuation(identify::package)
}
