% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_accent}
\alias{is_accent}
\title{Check the accent in vector}
\usage{
is_accent(x, type = "one")
}
\arguments{
\item{x}{not null class vector.}

\item{type}{character class vector that indicates the type of search in the vector. If "one" the search stops when true and does not check the entire vector. If "all" the search checks the entire vector.}
}
\description{
The \code{is_accent()} check if the vector has an accent. It's a function in experiment.
}
\examples{
is_accent("")                                    #FALSE
is_accent(".")                                   #FALSE
is_accent(" ")                                   #FALSE
is_accent("`")                                   #FALSE
is_accent("\\\\u00d7\\\\u00f7")                      #FALSE
is_accent("na")                                  #FALSE
is_accent("s\\\\u00e1")                            #TRUE
is_accent("aeiou")                               #FALSE
is_accent("aeiou uoiea")                         #FALSE
is_accent("AEIOU uoiea")                         #FALSE
is_accent(convert_utf8("\\\\u00e0\\\\u00e8\\\\u00ed")) #TRUE
is_accent(" a a ")                               #FALSE
is_accent(c(" a ", convert_utf8("\\\\u00e0")))     #TRUE
is_accent(convert_utf8("\\\\u00e7"))               #TRUE
is_accent("c")                                   #FALSE
is_accent(convert_utf8("\\\\u00e4"))               #TRUE
is_accent("a")                                   #FALSE
is_accent(convert_utf8("\\\\u00f5"))               #TRUE
is_accent(convert_utf8("\\\\u00d8"))               #TRUE
is_accent(" o ")                                 #FALSE
is_accent(NA)                                    #FALSE
is_accent(".5")                                  #FALSE
is_accent(convert_utf8("!#$\%\\\\u00a8&*(){]};."))  #FALSE
is_accent(5)                                     #FALSE
is_accent(c(1, NA, "na"))                        #FALSE
is_accent(c(56, NA, convert_utf8("\\\\u00d8")))    #TRUE

#FALSE
all(is_accent(c("", " ", "'", "/", "*", "-", "+", ".", ",", ";", "<", ">",
            ":", "!", "$", "\%", "&", "*", "(", ")", "_", "-",
            "{", "}", "[", "]", "=", "?", "1", "~", "`", "'")))
}
