% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_element}
\alias{is_element}
\title{Check the element in vector}
\usage{
is_element(x, type = "one")
}
\arguments{
\item{x}{not null class vector.}

\item{type}{character class vector that indicates the type of search in the vector. If "one" the search stops when true and does not check the entire vector. If "all" the search checks the entire vector.}
}
\description{
The \code{is_element()} check if the vector has an element.
}
\examples{
is_element(c("1", NA, "3"))          #TRUE
is_element(c(1, NA, 3))              #TRUE
is_element(c("", 
 convert_utf8("\\\\u00e7\\\\u00e7"), 
 convert_utf8("\\\\u00e1\\\\u00e9")))    #TRUE

# modify type
is_element(c(
 NA, 
 "", 
 convert_utf8("\\\\u00e7\\\\u00e7")), 
 type = "all")                       #FALSE FALSE TRUE
}
