% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_empty}
\alias{is_empty}
\title{Check the empty in vector}
\usage{
is_empty(x, type = "one")
}
\arguments{
\item{x}{not null class vector.}

\item{type}{character class vector that indicates the type of search in the vector. If "one" the search stops when true and does not check the entire vector. If "all" the search checks the entire vector.}
}
\description{
The \code{is_empty()} check if the vector has an empty.
}
\examples{
is_empty("")                                         #TRUE
is_empty(".")                                        #FALSE
is_empty(" ")                                        #FALSE
is_empty("  ")                                       #FALSE
is_empty("   ")                                      #FALSE
is_empty("na")                                       #FALSE
is_empty(NA)                                         #FALSE
is_empty(".5")                                       #FALSE
is_empty("!#$\%&*()_-[{]};.")                         #FALSE
is_empty(5)                                          #FALSE
is_empty(c("", NA_character_, "12mg"))               #TRUE
is_empty(c(NA_character_, NA_character_, "12 mg"))   #FALSE
}
