% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_letter}
\alias{is_letter}
\title{Check the letter in vector}
\usage{
is_letter(x, type = "one")
}
\arguments{
\item{x}{not null class vector.}

\item{type}{character class vector that indicates the type of search in the vector. If "one" the search stops when true and does not check the entire vector. If "all" the search checks the entire vector.}
}
\description{
The \code{is_letter()} check if the vector has an letter.
}
\examples{
is_letter("")                                #FALSE
is_letter(".")                               #FALSE
is_letter(" ")                               #FALSE
is_letter("na")                              #TRUE
is_letter("aeiou uoiea")                     #TRUE
is_letter("AEIOU uoiea")                     #TRUE
is_letter("AEIOU UOIEA")                     #TRUE
is_letter(convert_utf8("\\\\u00e1 \\\\u00f9"))   #FALSE
is_letter(convert_utf8("\\\\u00c1 \\\\u00e1"))   #FALSE
is_letter(NA)                                #FALSE
is_letter(".5")                              #FALSE
is_letter(5)                                 #FALSE
is_letter(c("1", NA, ""))                    #FALSE
is_letter(c("1", NA, "A"))                   #TRUE
is_letter(c(56, NA, "!#$\%&*()_-[{]};."))     #FALSE
is_letter(c(56, NA, "!#$\%&*()_-[{]};.5"))    #FALSE
is_letter(c(56, NA, "!#$\%&*(b)_-[{]};.5"))   #TRUE
}
