% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_punctuation}
\alias{is_punctuation}
\title{Check the punctuation in vector}
\usage{
is_punctuation(x, type = "one")
}
\arguments{
\item{x}{not null class vector.}

\item{type}{character class vector that indicates the type of search in the vector. If "one" the search stops when true and does not check the entire vector. If "all" the search checks the entire vector.}
}
\description{
The \code{is_punctuation()} check if the vector has an punctuation.
}
\examples{
is_punctuation(convert_utf8("s\\\\u00e1"))         #FALSE
is_punctuation("")                               #FALSE
is_punctuation(" ")                              #FALSE
is_punctuation("'")                              #TRUE
is_punctuation("/")                              #TRUE
is_punctuation("*")                              #TRUE
is_punctuation("-")                              #TRUE
is_punctuation("+")                              #TRUE
is_punctuation(".")                              #TRUE
is_punctuation(",")                              #TRUE
is_punctuation(";")                              #TRUE
is_punctuation("<")                              #TRUE
is_punctuation(">")                              #TRUE
is_punctuation(":")                              #TRUE
is_punctuation("!")                              #TRUE
is_punctuation("$")                              #FALSE
is_punctuation("\%")                              #TRUE
is_punctuation("&")                              #TRUE
is_punctuation("*")                              #TRUE
is_punctuation("(")                              #TRUE
is_punctuation(")")                              #TRUE
is_punctuation("_")                              #TRUE
is_punctuation("-")                              #TRUE
is_punctuation("{")                              #TRUE
is_punctuation("}")                              #TRUE
is_punctuation("]")                              #TRUE
is_punctuation("=")                              #TRUE
is_punctuation("?")                              #TRUE
is_punctuation("1")                              #FALSE
is_punctuation("~")                              #TRUE
is_punctuation(convert_utf8("\\\\u00e3"))          #FALSE
is_punctuation("a")                              #FALSE
is_punctuation(convert_utf8("\\\\u00b4"))          #TRUE
is_punctuation(convert_utf8("\\\\u00e1"))          #FALSE
is_punctuation("`")                              #TRUE
is_punctuation(convert_utf8("\\\\u00f9"))          #FALSE
is_punctuation("a")                              #FALSE
is_punctuation(" a")                             #FALSE
is_punctuation("a ")                             #FALSE
is_punctuation(" a ")                            #FALSE
is_punctuation(c(56, NA, "(246) 659 996 325"))   #TRUE
is_punctuation(c(56, NA, "$ 56,99"))             #TRUE
}
