% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_space}
\alias{is_space}
\title{Check the space in vector}
\usage{
is_space(x, type = "one")
}
\arguments{
\item{x}{not null class vector.}

\item{type}{character class vector that indicates the type of search in the vector. If "one" the search stops when true and does not check the entire vector. If "all" the search checks the entire vector.}
}
\description{
The \code{is_space()} check if the vector has an space.
}
\examples{
is_space("")                                 #FALSE
is_space(".")                                #FALSE
is_space(" ")                                #TRUE
is_space("  ")                               #TRUE
is_space("   ")                              #TRUE
is_space("na")                               #FALSE
is_space("na ")                              #TRUE
is_space(" na ")                             #TRUE
is_space("n a")                              #TRUE
is_space(NA)                                 #FALSE
is_space(".5")                               #FALSE
is_space("!#$\%&*()_-[{]};.")                 #FALSE
is_space(5)                                  #FALSE
is_space(c("10mg", NA_character_, "12mg"))   #FALSE
is_space(c("10mg", NA_character_, "12 mg"))  #TRUE
}
