% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{package}
\alias{package}
\title{Packages from language programming R}
\format{
A data frame with 8 rows and 3 columns:
\describe{
\item{package}{package name.}
\item{depends}{dependency from package.}
\item{license}{license of use from package.}
}
}
\source{
\code{rpkg::cran_package_all()}
}
\usage{
package
}
\description{
A dataset containing the package name, depends and license from packages from language programming R.
}
\keyword{datasets}
