% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stadium}
\alias{stadium}
\title{Stadiums in World}
\format{
A data frame with 12 rows and 3 columns:
\describe{
\item{country}{country name.}
\item{city}{city name and state name.}
\item{name}{stadium name.}
\item{club}{stadium to club.}
\item{capacity}{maximum capacity of stadium fans.}
}
}
\source{
\url{http://stadiumdb.com/}
}
\usage{
stadium
}
\description{
A dataset containing the the country, city name, stadium name and capacity from stadiums in World.
}
\keyword{datasets}
