% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{str_class}
\alias{str_class}
\title{Structure from column class in data frame}
\usage{
str_class(data, format_column = TRUE)
}
\arguments{
\item{data}{object class data frame.}

\item{format_column}{logical class vector to format column with characters of length > 16.}
}
\description{
The \code{str_class()} obtains information about the structure from column of the data frame.
}
\examples{
str_class(mtcars)
str_class(iris)
}
