% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{str_data}
\alias{str_data}
\title{Structure from data frame}
\usage{
str_data(data, measure = 'mb', format_column = TRUE)
}
\arguments{
\item{data}{object class data frame.}

\item{measure}{character class vector that indicates the unit of measure for calculating the object's size in the calling environment. Option is 'bytes', 'kb, 'mb', 'gb' and 'tb'.}

\item{format_column}{logical class vector to format column with characters of length > 16.}
}
\value{
o retorno e quadro de dados com informacao
}
\description{
The \code{str_data()} obtains information about the structure of the data frame.
}
\examples{
str_data(mtcars)
str_data(iris)
}
