% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{str_head}
\alias{str_head}
\title{Structure of column in data frame}
\usage{
str_head(data, format_column = TRUE)
}
\arguments{
\item{data}{object class data frame.}

\item{format_column}{logical class vector to format column with characters of length > 16.}
}
\description{
The \code{str_head()} is structure of the beginning of each column in the data frame.
}
\examples{
str_head(mtcars)
str_head(iris)
}
